--- index.js
@@ -1,954 +1,0 @@
-import express from 'express';
-import cors from 'cors';
-import dotenv from 'dotenv';
-import nodemailer from "nodemailer";
-import { MongoClient, ObjectId, ServerApiVersion } from 'mongodb';
-import cookieParser from 'cookie-parser';
-
-
-dotenv.config();
-const app = express();
-const port = process.env.PORT || 3000;
-
-// Middleware
-app.use(cookieParser());
-app.use(express.json());
-app.use(cors({
-    origin: ['http://localhost:5173',"https://wws-idp-website.vercel.app","https://graceful-sable-b6d5d1.netlify.app"],
-    credentials: true
-}));
-
-// Email transporter (Gmail Example)
-const transporter = nodemailer.createTransport({
-    service: "gmail",
-    auth: {
-        user: process.env.EMAIL_USER,   // তোমার Gmail
-        pass: process.env.EMAIL_PASS    // App Password (normal password নয়)
-    }
-});
-
-
-// MongoDB client
-const client = new MongoClient(process.env.MONGO_DB_URI, {
-    serverApi: {
-        version: ServerApiVersion.v1,
-        strict: true,
-        deprecationErrors: true,
-    },
-});
-
-// let dbCollections = {}; // All collections centralized
-
-async function run() {
-    try {
-        console.log("Connected to MongoDB");
-        const database = client.db('wwsDB');
-
-
-    const usersCollection = database.collection("users");
-    const helpCollection=database.collection("helpFrom")
-    const  coursesCollection=database.collection("courses")
-    const  scholarshipsCollection=database.collection("scholarships")
-    const universitiesCollection=database.collection("universities")
-    const eventsCollection= database.collection("events");
-    const collaborateCollection = database.collection("collaborate");
-    const popularCollection = database.collection("popular");
-
-
-
-
-    app.get("/popular",async(req,res)=>{
-
-    let result = await popularCollection.find().toArray()
-
-    res.send(result)
-
-
-    })
-
-
-
-    app.get("/popular/courses",async(req,res)=>{
-
-
-
-        const {
-        country,
-        level,
-        gpa_percent,
-        english_test,
-        total_score,
-        budget_per_year,
-        preferred_intake,
-        study_gaps_years
-    } = req.query;
-
-    let results =await popularCollection.find().toArray()
-
-    // 🔍 Filter: Country
-    if (country) {
-        results = results.filter(item =>
-            item.country.toLowerCase() === country.toLowerCase()
-        );
-    }
-
-    // 🔍 Filter: Level
-    if (level) {
-        results = results.filter(item =>
-            item.level.toLowerCase() === level.toLowerCase()
-        );
-    }
-
-    // 🔍 Filter: Minimum GPA (>=)
-    if (gpa_percent) {
-        results = results.filter(item =>
-            item.gpa_percent >= Number(gpa_percent)
-        );
-    }
-
-    // 🔍 Filter: English Test Type (IELTS / PTE / TOEFL)
-    if (english_test) {
-        results = results.filter(item =>
-            item.english_test.toLowerCase() === english_test.toLowerCase()
-        );
-    }
-
-    // 🔍 Filter: Total Score (>=)
-    if (total_score) {
-        results = results.filter(item =>
-            item.total_score >= Number(total_score)
-        );
-    }
-
-    // 🔍 Filter: Budget (<= max budget)
-    if (budget_per_year) {
-        results = results.filter(item =>
-            item.budget_per_year <= Number(budget_per_year)
-        );
-    }
-
-    // 🔍 Filter: Preferred Intake
-    if (preferred_intake) {
-        results = results.filter(item =>
-            item.preferred_intake.toLowerCase() === preferred_intake.toLowerCase()
-        );
-    }
-
-    // 🔍 Filter: Study Gap (<= allowed gap)
-    if (study_gaps_years) {
-        results = results.filter(item =>
-            item.study_gaps_years <= Number(study_gaps_years)
-        );
-    }
-
-    res.json({
-        success: true,
-        count: results.length,
-        data: results,
-    });
-
-    })
-
-       
-        app.get("/getUser/:email", async (req, res) => {
-
-            let email = req.params.email
-
-            let query = { email: email }
-            let result = await dbCollections.usersCollection.findOne(query)
-
-                        if (!result) {
-                            return res.send({ message: "No user found" })
-                        }
-                        let user = false
-                        if (result.role === "user") {
-                            user = true
-                        }
-
-                        res.send({ user })
-
-
-
-        })
-
-        app.get("/getAdmin/:email", async (req, res) => {
-
-            let email = req.params.email
-
-            let query = { email: email }
-            let result = await usersCollection.findOne(query)
-
-            if (!result) {
-                return res.send({ message: "No user found" })
-            }
-            let admin = false
-            if (result.role === "admin") {
-                admin = true
-            }
-
-            res.send({ admin })
-
-
-
-        })
-
-        app.get("/getAmbassador/:email", async (req, res) => {
-
-            let email = req.params.email
-
-            let query = { email: email }
-            let result = await usersCollection.findOne(query)
-
-            if (!result) {
-                return res.send({ message: "No user found" })
-            }
-            let ambassador = false
-            if (result.role === "ambassador") {
-                ambassador = true
-            }
-
-            res.send({ ambassador })
-
-
-
-        })
-
-        // ===== Ambassador Permission Routes =====
-        app.get('/user/ambassador', async (req, res) => {
-            try {
-                const result = await usersCollection.find({ role: "ambassador" }).toArray();
-                res.send(result);
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to fetch ambassadors' });
-            }
-        });
-
-         app.get("/ambassador/access/:email",async(req,res)=>{
-
-
-
-            try {
-                const { email } = req.params;
-
-                if (!email) {
-                return res.status(400).json({ message: "Email is required" });
-                }
-
-                let User=usersCollection
-
-                const user = await User.findOne(
-                        { email },
-                        { projection: { role: 1, scholarships: 1, courses: 1, universities: 1, events: 1 } }
-                        );
-
-                if (!user) {
-                return res.status(404).json({ message: "User not found" });
-                }
-
-                // ✅ role ambassador না হলে access deny
-                if (user.role !== "ambassador") {
-                return res.status(403).json({ message: "Access denied: Not an ambassador" });
-                }
-
-                // ✅ শুধু true field গুলো পাঠানো হবে
-                const trueFields = {};
-                if (user.scholarships) trueFields.scholarships = true;
-                if (user.courses) trueFields.courses = true;
-                if (user.universities) trueFields.universities = true;
-                if (user.events) trueFields.events = true;
-
-                return res.json({
-                role: user.role,
-                access: trueFields,
-                });
-            } catch (error) {
-                res.status(500).json({ message: error.message });
-            }
-
-        })
-
-        app.patch('/user/ambassador/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const permissions = req.body;
-
-                console.log(id,permissions)
-                
-                const query = ObjectId.isValid(id) ? { _id: new ObjectId(id) } : { _id: id };
-                
-                const updateDoc = {
-                    $set: permissions
-                };
-                
-                const result = await usersCollection.updateOne(query, updateDoc);
-                res.send({ success: true, result });
-            } catch (err) {
-                console.error('Permission update error:', err);
-                res.status(500).send({ success: false, message: 'Failed to update permissions' });
-            }
-        });
-
-        
-
-
-        app.post('/post-users', async (req, res) => {
-            try {
-                const user = req.body;
-                if (!user || !user.email) return res.status(400).send({ message: 'User data or email is missing' });
-                const existingUser = await usersCollection.findOne({ email: user.email });
-                if (existingUser) return res.status(409).send({ message: 'User already exists' });
-                const result = await usersCollection.insertOne(user);
-                res.status(201).send({ message: 'User added successfully', userId: result.insertedId });
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to add user' });
-            }
-        });
-
-        // ===== Help Routes =====
-        app.get('/help-from-wws', async (req, res) => {
-            try {
-                const result = await helpCollection.find().toArray();
-                res.send(result);
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to fetch enquiries' });
-            }
-        });
-
-
-        app.patch("/help-from-wws/:id", async (req, res) => {
-
-            let id = req.params.id
-            let status = req.body.status
-
-            let query = { _id: new ObjectId(id) }
-
-            let updatedDoc = {
-                $set: {
-                    status: status
-                }
-            }
-
-            let result = await helpCollection.updateOne(query, updatedDoc)
-
-            res.send(result)
-        });
-
-
-        app.get('/help-from-wws/:userEmail', async (req, res) => {
-            try {
-
-                let userEmail = req.params.userEmail
-
-                let query = { userEmail }
-                const result = await helpCollection.find(query).toArray();
-                res.send(result);
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to fetch enquiries' });
-            }
-        });
-
-
-        app.delete('/help-from-wws/:id', async (req, res) => {
-            try {
-                let id = req.params.id
-
-                let query = { _id: new ObjectId(id) }
-
-                const result = await helpCollection.deleteOne(query);
-                res.send(result);
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to delete enquiry' });
-            }
-        });
-
-        app.post('/help-from-wws', async (req, res) => {
-            try {
-                const enquiry = req.body;
-                if (!enquiry) return res.status(400).send({ message: 'No data provided' });
-                const result = await helpCollection.insertOne(enquiry);
-                res.send({ message: 'Enquiry submitted successfully', id: result.insertedId });
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to submit enquiry' });
-            }
-        });
-
-
-
-        app.get("/api/scholarships", async (req, res) => {
-            const data = await scholarshipsCollection.find().toArray();
-            res.send(data);
-
-
-        })
-
-        app.get("/api/universities", async (req, res) => {
-            const data = await universitiesCollection.find().toArray();
-            res.send(data);
-
-
-        })
-
-        app.get("/api/events", async (req, res) => {
-            const data = await eventsCollection.find().toArray();
-            res.send(data);
-
-
-        })
-
-        // ===== Search GET Routes (All) =====
-        app.get('/api/search/course', async (req, res) => {
-            try {
-                const data = await coursesCollection.find().toArray();
-                res.send({ success: true, data });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to fetch courses' });
-            }
-        });
-
-        app.get("/api/scholarship/:id", async (req, res) => {
-
-            let id = req.params.id
-
-            // OR query বানানো হচ্ছে
-            const query = {
-                $or: [
-                    { _id: id }, // string match
-                    ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null // ObjectId match (valid হলে)
-                ].filter(Boolean) // null বাদ দেওয়ার জন্য
-            };
-
-            let result = await scholarshipsCollection.findOne(query)
-
-            res.send(result)
-        })
-
-        app.get('/api/search/scholarships', async (req, res) => {
-            try {
-                const data = await scholarshipsCollection.find().toArray();
-                res.send({ success: true, data });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to fetch scholarships' });
-            }
-        });
-
-        app.get('/api/search/universities', async (req, res) => {
-            try {
-                const data = await universitiesCollection.find().toArray();
-                res.send({ success: true, data });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to fetch universities' });
-            }
-        });
-
-        app.get("/api/event/:id", async (req, res) => {
-
-            let id = req.params.id
-            console.log(id)
-
-            // OR query বানানো হচ্ছে
-            const query = {
-                $or: [
-                    { _id: id }, // string match
-                    ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null // ObjectId match (valid হলে)
-                ].filter(Boolean) // null বাদ দেওয়ার জন্য
-            };
-            let result = await eventsCollection.findOne(query)
-            console.log(result)
-
-            res.send(result)
-        })
-
-
-        app.get('/users', async (req, res) => {
-            try {
-                const data = await usersCollection.find().toArray();
-                res.send({ success: true, data });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to fetch users' });
-            }
-        });
-
-            // Update user (PATCH)
-        app.patch('/users/:id', async (req, res) => {
-        try {
-            const id = req.params.id;
-            const updateData = req.body;   // { role: 'admin' } or any field
-
-            const result = await usersCollection.updateOne(
-            { _id: new ObjectId(id) },
-            { $set: updateData }
-            );
-
-            if (result.modifiedCount === 0) {
-            return res.send({ success: false, message: 'No changes made or user not found' });
-            }
-
-            res.send({ success: true, message: 'User updated successfully' });
-        } catch (err) {
-            res.status(500).send({ success: false, message: 'Failed to update user' });
-        }
-        });
-
-
-        // Delete user (DELETE)
-        app.delete('/users/:id', async (req, res) => {
-        try {
-            const id = req.params.id;
-
-            const result = await usersCollection.deleteOne({ _id: new ObjectId(id) });
-
-            if (result.deletedCount === 0) {
-            return res.send({ success: false, message: 'User not found' });
-            }
-
-            res.send({ success: true, message: 'User deleted successfully' });
-        } catch (err) {
-            res.status(500).send({ success: false, message: 'Failed to delete user' });
-        }
-        });
-
-
-
-
-
-
-        app.get('/api/search/events', async (req, res) => {
-            try {
-                const data = await eventsCollection.find().toArray();
-                res.send({ success: true, data });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to fetch events' });
-            }
-        });
-
-        // ===== Scholarships Search POST Route (Fixed for flat structure) =====
-        app.post('/api/search/scholarships', async (req, res) => {
-            try {
-                const { studyLevel, destination } = req.body;
-                console.log("Scholarships Request body:", req.body);
-
-                // Build query for flat structure
-                const query = {};
-                if (studyLevel) query.studyLevel = { $regex: studyLevel, $options: "i" };
-                if (destination) query.destination = { $regex: destination, $options: "i" };
-
-                console.log("Scholarships query:", query);
-
-                // Find matching documents
-                const results = await scholarshipsCollection.find(query).toArray();
-                console.log("Scholarships results:", results);
-
-                res.json({ success: true, data: results });
-            } catch (error) {
-                console.error("Scholarships search error:", error);
-                res.status(500).json({ success: false, message: "Scholarship search failed" });
-            }
-        });
-
-
-
-
-        app.get("/api/university/:id", async (req, res) => {
-
-            let id = req.params.id
-
-            // OR query বানানো হচ্ছে
-            const query = {
-                $or: [
-                    { _id: id }, // string match
-                    ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null // ObjectId match (valid হলে)
-                ].filter(Boolean) // null বাদ দেওয়ার জন্য
-            };
-
-            let result=await dbCollections.universitiesCollection.findOne(query)
-
-            res.send(result)
-        })
-
-        // ===== Universities Search POST Route (Fixed for flat structure) =====
-        app.post('/api/search/universities', async (req, res) => {
-            try {
-                const { universityName, destination } = req.body;
-                console.log("Universities Request body:", req.body);
-
-                // Build query for flat structure
-                const query = {};
-                if (universityName) query.universityName = { $regex: universityName, $options: "i" };
-                if (destination) query.destination = { $regex: destination, $options: "i" };
-
-                console.log("Universities query:", query);
-
-                // Find matching documents
-                const results = await universitiesCollection.find(query).toArray();
-                console.log("Universities results:", results);
-
-                res.json({ success: true, data: results });
-            } catch (error) {
-                console.error("Universities search error:", error);
-                res.status(500).json({ success: false, message: "University search failed" });
-            }
-        });
-
-        // ===== Events Search POST Route (Fixed for flat structure) =====
-        app.post('/api/search/events', async (req, res) => {
-            try {
-                const { city, month, destination } = req.body;
-                console.log("Events Request body:", req.body);
-
-                // Build query for flat structure
-                const query = {};
-                if (city) query.city = { $regex: city, $options: "i" };
-                if (month) query.month = { $regex: month, $options: "i" };
-                if (destination) query.destination = { $regex: destination, $options: "i" };
-
-                console.log("Events query:", query);
-
-                // Find matching documents
-                const results = await eventsCollection.find(query).toArray();
-                console.log("Events results:", results);
-
-                res.json({ success: true, data: results });
-            } catch (error) {
-                console.error("Events search error:", error);
-                res.status(500).json({ success: false, message: "Event search failed" });
-            }
-        });
-
-
-        app.get("/api/course/:id", async (req, res) => {
-
-
-            let id = req.params.id
-            // console.log(id)
-
-            // OR query বানানো হচ্ছে
-            const query = {
-                $or: [
-                    { _id: id }, // string match
-                    ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null // ObjectId match (valid হলে)
-                ].filter(Boolean) // null বাদ দেওয়ার জন্য
-            };
-
-            let result = await coursesCollection.findOne(query)
-
-            res.send(result)
-        })
-
-        // ===== Course Search POST Route (Fixed for flat structure) =====
-        app.post("/api/search/course", async (req, res) => {
-            try {
-                const { subject, studyLevel, destination } = req.body;
-                console.log("Course Request body:", req.body);
-
-                // Build query for flat structure
-                const query = {};
-                if (subject) query.subject = { $regex: subject, $options: "i" };
-                if (studyLevel) query.studyLevel = { $regex: studyLevel, $options: "i" };
-                if (destination) query.destination = { $regex: destination, $options: "i" };
-
-                console.log("Course query:", query);
-
-                // Find matching documents
-                const results = await coursesCollection.find(query).toArray();
-                console.log("Course results:", results);
-
-                res.json({ success: true, data: results });
-            } catch (error) {
-                console.error("Course search error:", error);
-                res.status(500).json({ success: false, message: "Course search failed" });
-            }
-        });
-
-        app.get('/collaborate', async (req, res) => {
-            try {
-                const data = await collaborateCollection.find().toArray();
-                res.send(data);
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to fetch universities' });
-            }
-        });
-
-        app.post('/collaborate', async (req, res) => {
-            try {
-                let email = "abdshakaet@gmail.com"
-                const enquiry = req.body;
-                // console.log(enquiry)
-                if (!enquiry) return res.status(400).send({ message: 'No data provided' });
-                const result = await collaborateCollection.insertOne(enquiry);
-                res.send({ message: 'Enquiry submitted successfully', id: result.insertedId });
-                // 2. Email পাঠাও
-                // const mailOptions = {
-                // from: process.env.EMAIL_USER,
-                // to: email,   // ইউজারের original email
-                // subject: "Your Post has been Submitted",
-                // text: `${enquiry.email}, your post has been submitted successfully. We will contact you soon!`
-                // };
-
-                const mailOptions = {
-                    from: process.env.EMAIL_USER,
-                    to: email,   // তোমার ইমেইল, যেটাতে message যাবে
-                    subject: "New Post Submitted by User",
-                    text: `
-                    Hi Abdulla Al Shakaet,
-
-                    A new post has been submitted by a user. Here are the details:
-
-                    ----------------------------------------
-                    Name: ${enquiry.name}
-                    Email: ${enquiry.email}
-                    Message: ${enquiry.message}
-                    ----------------------------------------
-
-                    Please follow up with the user as needed.
-
-                    Thank you,
-                    World Wise Scholar Team
-                    `
-                };
-
-
-                await transporter.sendMail(mailOptions);
-
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to submit enquiry' });
-            }
-        });
-
-
-
-
-        app.post('/add-new-scholarship', async (req, res) => {
-            try {
-                const data = req.body;
-
-
-                const result = await scholarshipsCollection.insertOne(data);
-                res.status(201).send({ message: 'Scholarship added successfully', userId: result.insertedId });
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to add Scholarship' });
-            }
-        });
-
-        // Normal Update Scholarship by id
-        app.put('/api/scholarship/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                // OR query বানানো হচ্ছে
-                const query = {
-                    $or: [
-                        { _id: id }, // string match
-                        ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null // ObjectId match (valid হলে)
-                    ].filter(Boolean) // null বাদ দেওয়ার জন্য
-                };
-
-                const updateDoc = {
-                    $set: req.body   // সরাসরি যা আসবে body থেকে, সেটাই update হবে
-                };
-
-                const result = await scholarshipsCollection.updateOne(query, updateDoc);
-
-                res.send(result);   // result এর মধ্যে matchedCount, modifiedCount থাকবে
-            } catch (error) {
-                console.error(error);
-                res.status(500).send({ message: 'Error updating scholarship' });
-            }
-        });
-
-
-        // Delete a scholarship by id
-        app.delete('/api/scholarship/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const query = ObjectId.isValid(id) ? { _id: new ObjectId(id) } : { _id: id };
-                const result = await scholarshipsCollection.deleteOne(query);
-                res.send({ success: true, deletedCount: result.deletedCount });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to delete scholarship' });
-            }
-        });
-
-        app.post('/add-new-university', async (req, res) => {
-            try {
-                const data = req.body;
-
-
-                const result = await universitiesCollection.insertOne(data);
-                res.status(201).send({ message: 'University added successfully', userId: result.insertedId });
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to add University' });
-            }
-        });
-
-        // Normal Update University by id (pattern aligned with scholarship)
-        app.put('/api/university/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const query = {
-                    $or: [
-                        { _id: id },
-                        ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null
-                    ].filter(Boolean)
-                };
-
-                const updateDoc = { $set: req.body };
-                const result = await universitiesCollection.updateOne(query, updateDoc);
-                res.send(result);
-            } catch (error) {
-                console.error(error);
-                res.status(500).send({ message: 'Error updating university' });
-            }
-        });
-
-        // Delete a university by id
-        app.delete('/api/university/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const query = ObjectId.isValid(id) ? { _id: new ObjectId(id) } : { _id: id };
-                const result = await universitiesCollection.deleteOne(query);
-                res.send({ success: true, deletedCount: result.deletedCount });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to delete university' });
-            }
-        });
-
-        // ==================Get All Course ==================
-        app.get("/api/course", async (req, res) => {
-            const data = await coursesCollection.find().toArray();
-            res.send(data);
-        })
- 
-        // ================== Add New Course ==================
-        app.post('/add-new-course', async (req, res) => {
-            try {
-                const data = req.body;
-
-                const result = await coursesCollection.insertOne(data);
-                res.status(201).send({ message: 'Course added successfully', courseId: result.insertedId });
-            } catch (err) {
-                res.status(500).send({ message: 'Failed to add Course' });
-            }
-        });
-
-        // ================== Update Course by ID ==================
-        app.put('/api/course/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const query = {
-                    $or: [
-                        { _id: id },
-                        ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null
-                    ].filter(Boolean)
-                };
-
-                const updateDoc = { $set: req.body };
-                const result = await coursesCollection.updateOne(query, updateDoc);
-                res.send(result);
-            } catch (error) {
-                console.error(error);
-                res.status(500).send({ message: 'Error updating course' });
-            }
-        });
-
-        // ================== Delete Course by ID ==================
-        app.delete('/api/course/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const query = ObjectId.isValid(id) ? { _id: new ObjectId(id) } : { _id: id };
-                const result = await coursesCollection.deleteOne(query);
-                res.send({ success: true, deletedCount: result.deletedCount });
-            } catch (err) {
-                res.status(500).send({ success: false, message: 'Failed to delete course' });
-            }
-        });
-
-
-        // ================== Add New Event ==================
-        app.post('/add-new-event', async (req, res) => {
-            try {
-                const data = req.body;
-                const result = await eventsCollection.insertOne(data);
-                res.status(201).send({ message: 'Event added successfully', eventId: result.insertedId });
-            } catch (err) {
-                console.error(err);
-                res.status(500).send({ message: 'Failed to add Event' });
-            }
-        });
-
-        // ================== Update Event by ID ==================
-        app.put('/api/event/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-
-                const query = {
-                    $or: [
-                        { _id: id },
-                        ObjectId.isValid(id) ? { _id: new ObjectId(id) } : null
-                    ].filter(Boolean)
-                };
-
-                const updateDoc = { $set: req.body };
-                const result = await eventsCollection.updateOne(query, updateDoc);
-                res.send(result);
-            } catch (error) {
-                console.error(error);
-                res.status(500).send({ message: 'Error updating event' });
-            }
-        });
-
-        // ================== Delete Event by ID ==================
-        app.delete('/api/event/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const query = ObjectId.isValid(id) ? { _id: new ObjectId(id) } : { _id: id };
-                const result = await eventsCollection.deleteOne(query);
-                res.send({ success: true, deletedCount: result.deletedCount });
-            } catch (err) {
-                console.error(err);
-                res.status(500).send({ success: false, message: 'Failed to delete event' });
-            }
-        });
-
-        // // ================== Get All Events ==================
-        // app.get('/api/events', async (req, res) => {
-        //     try {
-        //         const events = await dbCollections.eventsCollection.find().toArray();
-        //         res.send({ success: true, data: events });
-        //     } catch (err) {
-        //         console.error(err);
-        //         res.status(500).send({ success: false, message: 'Failed to fetch events' });
-        //     }
-        // });
-
-        // ================== Get Single Event by ID ==================
-        app.get('/api/event/:id', async (req, res) => {
-            try {
-                const id = req.params.id;
-                const query = ObjectId.isValid(id) ? { _id: new ObjectId(id) } : { _id: id };
-                const event = await eventsCollection.findOne(query);
-
-                if (!event) {
-                    return res.status(404).send({ success: false, message: 'Event not found' });
-                }
-
-                res.send({ success: true, data: event });
-            } catch (err) {
-                console.error(err);
-                res.status(500).send({ success: false, message: 'Failed to fetch event' });
-            }
-        });
-
-
-    } catch (err) {
-        console.error(err);
-    }
-}
-
-run().catch(console.dir);
-
-app.get('/', (req, res) => {
-    res.send("World Wise Scholar Server is cooking...!");
-});
-
-app.listen(port, () => {
-    console.log('Server running on port', port);
-});
