+++ package.json
@@ -4,6 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "start": "node dist/app.js",
+    "dev": "ts-node-dev --respawn --transpile-only app.ts",
+    "build": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -17,5 +20,14 @@
     "express": "^5.1.0",
     "mongodb": "^6.19.0",
     "nodemailer": "^7.0.6"
+  }
+  ,"devDependencies":{
+    "typescript":"^5.6.2",
+    "ts-node-dev":"^2.0.0",
+    "@types/express":"^4.17.17",
+    "@types/node":"^20.5.1",
+    "@types/cookie-parser":"^1.4.3",
+    "@types/cors":"^2.8.17",
+    "@types/nodemailer":"^6.4.7"
   }
 }
